Added new functions. Spaw function added which is really just the buy function from the regular shop. TSAs aren't charged when buying items however so it can be used effectively to conveniently spawn a large number of any item. (Beware if you're a normal user - hasn't been tested but you'll probably be charged the tc cost)
Get inventory added. Just returns a list of all items in a user's inventory. Useful to take a snapshot of the inventory at time point A, generate a load of items or something and then take another snapshot now at time point B. All additional items will be able to be seen as the difference between snapshots and you can easily get information about just the newest or changed items.